Getting ready to test vegetation model with zarr chunk stats comparison.
